pyproject.toml
[tool.poetry]
name = "cdk-deploy-ec2"
version = "0.1.1"
description = "Deploy EC2 instances using AWS CDK (Python)"
authors = ["TimekillerTK <erwartungen@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.93.0"
requests = "^2.27.1"
urllib3 = "^2.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "strict"

exclude = [
    "**/node_modules",
    "**/__pycache__",
    "venv/",
    ".venv/"
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 100

[tool.ruff]
# E & F - defaults
# W     - optional warnings (for example trailing whitespace)
# C90   - code complexity
# ANN   - enforce strict type annotation
select = ["E", "F", "W", "C90", "ANN"]
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
]